tests/submission/Infraware/Session History/contents/Session_History/Not Use/location_reload.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>location_reload</h3>
		<li>Click 'reload' button.</li>
	</p>
	
	<hr>

	<div id="log"></div>

	<iframe id="frame" style="width: 0px; height: 0px;"></iframe>
	<button onclick='onClickBtn()'>reload</button>
	
	<script>
		var t1 = async_test('location reload test');
		
		iframeWindow = document.getElementById('frame').contentWindow;
		var framestr = iframeWindow.localStorage.getItem("load");
		
		function onClickBtn() {
			if (('location' in window) && window.location != null
				&& ('location' in document) && document.location != null){

				var len = window.history.length;
				var href = location.href;
				
				iframeWindow.localStorage.setItem("load", "true");
				location.reload();

			} else {
				test(function() {
					assert_unreached("location interface is not supported.");
				});
			}		
		}

		function onLoadEvent(e) {
			if (framestr != "true") return;
			
			t1.step(function() {
				assert_true(true, "load event is raised.");
			});
			t1.done();
			
			iframeWindow.localStorage.setItem("load", "false");
		}

		if (window.addEventListener) {
            window.addEventListener('load', onLoadEvent, false);
        }
        else if (window.attachEvent) {
            window.attachEvent("onload", onLoadEvent);
        }
        else {
            test(function() {
				assert_unreached("DOM event subscribtion is not supported.");
			});
        }

	</script>
 </body>
</html>
